Crate quicr_core

Structs

ApplicationClose

Config
Parameters governing the core QUIC state machine.

ConnectionClose

ConnectionHandle

ConnectionId
Protocol-level identifier for a connection.

Endpoint
The main entry point to the library

ListenConfig
Information needed to accept incoming connections.

NewConnection
Information about a new incoming connection

PersistentState
Information that should be preserved between restarts.

StreamId
Identifier for a stream within a particular connection

TransportError

Enums

ConnectionError
Reasons why a connection might be lost.

Directionality
Whether a stream communicates data in both directions or only from the initiator

Event
Events of interest to the application

Io
I/O operations to be immediately executed the backend.

ReadError

Side
Whether an endpoint was the initiator of a connection

Timer

WriteError

Constants

VERSION
The QUIC protocol version implemented