Enum quicr_core::Event

pub enum Event {
    Connected {
        protocol: Option<Box<[u8]>>,
    },
    ConnectionLost {
        reason: ConnectionError,
    },
    ConnectionDrained,
    StreamReadable {
        stream: StreamId,
        fresh: bool,
    },
    StreamWritable {
        stream: StreamId,
    },
    StreamFinished {
        stream: StreamId,
    },
    StreamAvailable {
        directionality: Directionality,
    },
}

Events of interest to the application

Variants
Connected

A connection was successfully established.

Fields of Connected

protocol: Option<Box<[u8]>>

ConnectionLost

A connection was lost.

Fields of ConnectionLost

reason: ConnectionError

ConnectionDrained

A closed connection was dropped.

StreamReadable

A stream has data or errors waiting to be read

Fields of StreamReadable

stream: StreamId

The affected stream

fresh: bool

Whether this is the first event on the stream

StreamWritable

A formerly write-blocked stream might now accept a write

Fields of StreamWritable

stream: StreamId

StreamFinished

All data sent on stream has been received by the peer

Fields of StreamFinished

stream: StreamId

StreamAvailable

At least one new stream of a certain directionality may be opened

Fields of StreamAvailable

directionality: Directionality