Struct quicr_core::TransportError

pub struct TransportError(_);

Methods

impl Error

pub fn frame(ty: Type) -> Self

impl Error

pub const NO_ERROR: Self

NO_ERROR: Self = Error(0)

the connection is being closed abruptly in the absence of any error

pub const INTERNAL_ERROR: Self

INTERNAL_ERROR: Self = Error(1)

the endpoint encountered an internal error and cannot continue with the connection

pub const SERVER_BUSY: Self

SERVER_BUSY: Self = Error(2)

the server is currently busy and does not accept any new connections

pub const FLOW_CONTROL_ERROR: Self

FLOW_CONTROL_ERROR: Self = Error(3)

an endpoint received more data than it permitted in its advertised data limits

pub const STREAM_ID_ERROR: Self

STREAM_ID_ERROR: Self = Error(4)

an endpoint received a frame for a stream identifier that exceeded its advertised maximum stream ID

pub const STREAM_STATE_ERROR: Self

STREAM_STATE_ERROR: Self = Error(5)

an endpoint received a frame for a stream that was not in a state that permitted that frame

pub const FINAL_OFFSET_ERROR: Self

FINAL_OFFSET_ERROR: Self = Error(6)

an endpoint received a STREAM frame containing data that exceeded the previously established final offset, or a RST_STREAM frame containing a final offset that was lower than the maximum offset of data that was already received, or a RST_STREAM frame containing a different final offset to the one already established

pub const FRAME_FORMAT_ERROR: Self

FRAME_FORMAT_ERROR: Self = Error(7)

an endpoint received a frame that was badly formatted. For instance, an empty STREAM frame that omitted the FIN flag, or an ACK frame that has more acknowledgment ranges than the remainder of the packet could carry

pub const TRANSPORT_PARAMETER_ERROR: Self

TRANSPORT_PARAMETER_ERROR: Self = Error(8)

an endpoint received transport parameters that were badly formatted, included an invalid value, was absent even though it is mandatory, was present though it is forbidden, or is otherwise in error

pub const VERSION_NEGOTIATION_ERROR: Self

VERSION_NEGOTIATION_ERROR: Self = Error(9)

an endpoint received transport parameters that contained version negotiation parameters that disagreed with the version negotiation that it performed, constituting a potential version downgrade attack

pub const PROTOCOL_VIOLATION: Self

PROTOCOL_VIOLATION: Self = Error(10)

an endpoint detected an error with protocol compliance that was not covered by more specific error codes

pub const UNSOLICITED_PATH_RESPONSE: Self

UNSOLICITED_PATH_RESPONSE: Self = Error(11)

an endpoint received a PATH_RESPONSE frame that did not correspond to any PATH_CHALLENGE frame that it previously sent

pub const TLS_HANDSHAKE_FAILED: Self

TLS_HANDSHAKE_FAILED: Self = Error(513)

the TLS handshake failed

pub const TLS_FATAL_ALERT_GENERATED: Self

TLS_FATAL_ALERT_GENERATED: Self = Error(514)

a TLS fatal alert was sent, causing the TLS connection to end prematurely

pub const TLS_FATAL_ALERT_RECEIVED: Self

TLS_FATAL_ALERT_RECEIVED: Self = Error(515)

a TLS fatal alert was received, causing the TLS connection to end prematurely